FEAT: Monothreaded mutability of the grid during transit
This is cool!  I can now scan the entire grid, mutating it as I
go along.  This means that I can now apply the single-threaded
features of PamSeam directly, just writing the algorith and letting
these exemplary tools do the horizontal and vertical scanning,
and concentrate only on the working algorithm, whichever one I'm
currently working on.  I don't know how this will interact well
with the "Scan all your neighbors" thinking of the original,
and I can already feel in the back of my mind a few shortcomings
to this approach, nevermind the trick of multi-threading the
damn thing.