FEAT Seam validation
I've been trying to figure out how to determine that a seam is "correct";
that it fits a dimension of the lattice, that it goes from one end of the
lattice to the other across that dimension, and that it never doubles back
or goes in the wrong direction.  This algorithm asserts that.